Fix type error, improve CI pipeline, and complete example coverage
- Fix TS2339 type error in check-spec-schema-sync.ts with proper
  type assertion for allOf conditional array
- Bump ajv 8.17.1->8.18.0 and @types/node 20.19.30->20.19.33
- Add spec/** path trigger to CI so spec changes run validation
- Add tsc --noEmit, check:refs, and check:coverage to CI pipeline
- Extract shared AJV utilities into scripts/lib/ajv-utils.ts to
  remove duplication between validate-schemas.ts and validate-examples.ts
- Add annotations, asset-index, and provenance example files to
  achieve 15/15 schema coverage (was 12/15)
