+++ examples/comprehensive-document/assets/index.json
@@ -1,0 +1,111 @@
+{
+  "version": "0.1",
+  "families": [
+    {
+      "name": "Inter",
+      "fonts": [
+        {
+          "id": "font-inter-regular",
+          "weight": 400,
+          "style": "normal"
+        },
+        {
+          "id": "font-inter-bold",
+          "weight": 700,
+          "style": "normal"
+        },
+        {
+          "id": "font-inter-italic",
+          "weight": 400,
+          "style": "italic"
+        }
+      ]
+    }
+  ],
+  "assets": [
+    {
+      "id": "hero-image",
+      "path": "images/hero.png",
+      "type": "image/png",
+      "size": 245760,
+      "hash": "sha256:a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2",
+      "metadata": {
+        "width": 1920,
+        "height": 1080,
+        "colorSpace": "sRGB",
+        "hasAlpha": false,
+        "dpi": 72
+      },
+      "license": {
+        "name": "CC BY 4.0",
+        "url": "https://creativecommons.org/licenses/by/4.0/"
+      },
+      "variants": [
+        {
+          "path": "images/hero-thumb.png",
+          "width": 480,
+          "size": 20480
+        },
+        {
+          "path": "images/hero-medium.png",
+          "width": 960,
+          "size": 81920
+        }
+      ]
+    },
+    {
+      "id": "diagram-svg",
+      "path": "images/architecture.svg",
+      "type": "image/svg+xml",
+      "size": 8192,
+      "hash": "sha256:b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3"
+    },
+    {
+      "id": "font-inter-regular",
+      "path": "fonts/Inter-Regular.woff2",
+      "type": "font/woff2",
+      "size": 98304,
+      "hash": "sha256:c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4",
+      "metadata": {
+        "family": "Inter",
+        "weight": 400,
+        "style": "normal"
+      }
+    },
+    {
+      "id": "font-inter-bold",
+      "path": "fonts/Inter-Bold.woff2",
+      "type": "font/woff2",
+      "size": 102400,
+      "hash": "sha256:d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5",
+      "metadata": {
+        "family": "Inter",
+        "weight": 700,
+        "style": "normal"
+      }
+    },
+    {
+      "id": "font-inter-italic",
+      "path": "fonts/Inter-Italic.woff2",
+      "type": "font/woff2",
+      "size": 100352,
+      "hash": "sha256:e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6",
+      "metadata": {
+        "family": "Inter",
+        "weight": 400,
+        "style": "italic"
+      }
+    },
+    {
+      "id": "supplementary-pdf",
+      "path": "embeds/appendix-data.pdf",
+      "type": "application/pdf",
+      "size": 524288,
+      "hash": "sha256:f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1b2c3d4e5f6a1",
+      "metadata": {
+        "filename": "appendix-data.pdf",
+        "description": "Supplementary data tables"
+      }
+    }
+  ]
+}
