+++ examples/signed-document/security/annotations.json
@@ -1,0 +1,38 @@
+{
+  "version": "0.1",
+  "annotations": [
+    {
+      "id": "ann-001",
+      "type": "comment",
+      "anchor": {
+        "blockId": "para-1",
+        "start": 0,
+        "end": 45
+      },
+      "author": "Jane Reviewer",
+      "created": "2025-01-12T09:15:00Z",
+      "content": "This section needs a stronger opening statement."
+    },
+    {
+      "id": "ann-002",
+      "type": "highlight",
+      "anchor": {
+        "blockId": "para-3"
+      },
+      "author": "John Editor",
+      "created": "2025-01-13T11:30:00Z",
+      "content": "Key conclusion paragraph"
+    },
+    {
+      "id": "ann-003",
+      "type": "note",
+      "anchor": {
+        "blockId": "heading-2",
+        "offset": 0
+      },
+      "author": "Jane Reviewer",
+      "created": "2025-01-14T08:00:00Z",
+      "content": "Consider restructuring this section before signing."
+    }
+  ]
+}
